removing nulls (consistency)
