+++ opsml/app/static/src/routes/opsml/model/card/versions/+page.ts
@@ -1,25 +1,29 @@
-import { getRegistryPage } from "$lib/scripts/registry_page";
+import { getRegistryPage } from "$lib/scripts/utils";
 import { type CardRequest, type CardResponse } from "$lib/scripts/types";
 import { listCards } from "$lib/scripts/utils";
 
 /** @type {import('./$types').PageLoad} */
 export async function load({ fetch, params, url }) {
-  const name: string = (url as URL).searchParams.get("name");
-  const repository: string = (url as URL).searchParams.get("repository");
-  const registry: string = (url as URL).searchParams.get("registry");
+  const name = (url as URL).searchParams.get("name") as string | undefined;
+  const repository = (url as URL).searchParams.get("repository") as
+    | string
+    | undefined;
+  const registry = (url as URL).searchParams.get("registry") as
+    | string
+    | undefined;
 
   const registryPage = await getRegistryPage(
-    registry,
+    registry!,
     undefined,
-    repository,
-    name,
+    repository as string | undefined,
+    name as string | undefined,
     0
   );
 
   const cardReq: CardRequest = {
     name,
     repository,
-    registry_type: registry,
+    registry_type: registry!,
     page: 0,
   };
