+++ opsml/app/static/src/routes/opsml/run/+page.svelte
@@ -4,7 +4,7 @@
   import {
   getRegistryPage,
   getRegistryStats,
-} from "$lib/scripts/registry_page";
+} from "$lib/scripts/utils";
   import { type PaginationSettings } from '@skeletonlabs/skeleton';
   import {
   type registryStats,
